requirements.txt
anyio==4.2.0
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.6
pillow==10.2.0
python-telegram-bot==21.1.1
PyYAML==6.0.1
sniffio==1.3.0
